The full item view page looks almost the same as in the v5.
