+++ src/app/item-page/clarin-files-section/clarin-files-section.component.scss
@@ -1,0 +1,3 @@
+/**
+The styling file for the `clarin-files-section.component`
+ */
