+++ src/app/item-page/simple/item-page.component.scss
@@ -21,7 +21,7 @@
 }
 
 .clarin-cut-bottom-2 {
-  margin-bottom: -24px;
+  margin-bottom: -18px;
 }
 
 .clarin-cut-top-1 {
