+++ src/assets/i18n/en.json5
@@ -2655,7 +2655,7 @@
 
   "item.page.link.full": "Show full item record",
 
-  "item.page.link.simple": "Simple item page",
+  "item.page.link.simple": "Show simple item record",
 
   "item.page.orcid.title": "ORCID",
